factors out tests
