+++ src/lib.rs
@@ -7,6 +7,7 @@ extern crate socool_parser;
 #[macro_use]
 extern crate serde_derive;
 extern crate num_rational;
+extern crate rayon;
 
 #[macro_use]
 pub mod macros;
